US12175 DT8545 Field labels with "less than or equal to" and "greater than or equal to" characters don't show in PDF (2) Add in gitignore file so that Downloads is source controlled. Also added in a couple lines in TestContext.cs so that the gitignore won't be deleted every run

+++ Medidata.RBT/TestContext.cs
@@ -338,7 +338,8 @@ namespace Medidata.RBT
         {
             List<String> createdFiles = Directory.GetFiles(TestContext.DownloadPath).ToList();
             foreach (String filePath in createdFiles)
-                File.Delete(filePath);
+                if(!filePath.EndsWith(".gitignore"))
+                    File.Delete(filePath);
         }
 
         /// <summary>
@@ -363,9 +364,8 @@ namespace Medidata.RBT
             {
                 DateTime lastWriteTime = File.GetLastWriteTime(filePath);
                 if (lastWriteTime >= CurrentFeatureStartTime)
-                {
-                    File.Delete(filePath);
-                }
+                    if (!filePath.EndsWith(".gitignore"))
+                        File.Delete(filePath);
             }
         }
